[IMPROVEMENT] Added Peer Detection, Skeleton code
Added ability for the peers to communicate and detect each other.
Currently, peers talk to all peers but it is relatively trivial to make
them talk directly.
Skeleton code has been added to maintain a list of peers, which can then
be used for communicating directly with other peers in the network.

Also added the skeleton code for communicating commands over the
network. This still needs some work, the initial idea was to have a
parent command object which declared methods for converting to and from
the network, which child objects could then implement (but this
immediately becomes a bad idea when you realise a child object has to
convert the object into itself, a loopy loop of logic). So, the next
idea was to store name/objects in a dictionary, send the name, look it
up in the dictionary then cast the data into that object. but this then
seemed like a dangerous idea, and needless. storing the name/objects
seemed like a good idea, but all that needs to be done is sending the
data of the object in a payload, then mapping that data to an instant of
the object on our end. Basically, all that needs to be sent is a name of
an object, and a payload which maps to data in the object.
The user can then operate on this object. Basically, replicating the
object that is sent from one peer on another, and responding to it. This
also raises the idea that instead of using a delegate to process
responses to actions, the responses could be baked into the command
objects themselves. Then once they are replicated, the response action
can be called immediately, operating on the data provided.
From a users perspective, this would mean all they need to do is create
commands and register them.