[IMPROVEMENT] Commands communicated with JSON
Commands are now send to peers as json, serialised before sending and
deserialised on receipt.

This introduces the dependency on Json.NET, but makes it easier to send
more complex commands without the hassle of parsing a custom protocol.
Also allows them to be more flexible.

Looks like it will require more refactoring, currently handling what to
do with the responses is going towards a long network response function.
It would be nicer to tidy this up and encapsulate responses into the
command objects themselves. Will probably remove the heirarchy currently
existing for commands. One standard library command which has a name and
a payload. Then the payload can be a json serialised version of the
command which can be loaded into a specific command which has a
PerformAction function, encapsulating the response.