[IMPROVEMENT] Added payloads to commands
Commands now have payloads which map to payload objects. Payloads are
serialised into json and attached to commands as networkPayloads just
before sending. Then they can be deserialised back into payload objects
for use in PerformAction. This is currently left to the user to do. Next
I'd like to make deserialisation automatic upon receipt so the user
doesn't have to do this and can simply access the payload data when
using it in perform action.