[IMPROVEMENT] dynamic payloads
Payloads are now dynamic, so the user can access properties they're
expecting and cast them to the types they need. This removes the need
for Payload types, and removes the need for keeping a record of payload
types. It also gets rid of the problem where a user had to deserialise
the network payload themselves (also this is still an available option
if they want the typing back).

Overall, it reduces the amount of code necessary to write, receieve and
operate on commands and their payloads.

Now, the flow looks something like this:
1. Create Command, give it a name and Define it's Perform Action.
2. Register it with the network

Now the peer can receive commands of that type and act accordingly

Sending commands:
1. Create Command, give it a name and Define it's Perform Action. Or
re-use command already created.
2.Supply Payload (can be done via anonymous type)
3. Pass it to the networks sendall method